refactor(infrastructure): move storage adapters into infrastructure layer
