+++ packages/core/src/infrastructure/storage/storage-service.ts
@@ -1,0 +1,8 @@
+import type { Buffer } from 'node:buffer';
+
+export interface StorageService {
+  uploadFile(path: string, file: Buffer): Promise<string>;
+  getFile(path: string): Promise<Buffer>;
+  deleteFile(path: string): Promise<void>;
+  getFileUrl(path: string): Promise<string>;
+}
