+++ packages/core/src/storage/storage-service.ts
@@ -1,8 +1,1 @@
-import type { Buffer } from "node:buffer";
-
-export interface StorageService {
-	uploadFile(path: string, file: Buffer): Promise<string>;
-	getFile(path: string): Promise<Buffer>;
-	deleteFile(path: string): Promise<void>;
-	getFileUrl(path: string): Promise<string>;
-}
+export * from '../infrastructure/storage/storage-service.js';
